+++ packages/angular-workspace/example-client-app/src/app/customapp/fv/fv-section.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { OkFvAccordionItemModule } from '@ni/ok-angular/fv/accordion-item';
+import { FvAccordionItemSectionComponent } from './fv-accordion-item-section.component';
+import { FvSectionComponent } from './fv-section.component';
+import { SubContainerModule } from '../sub-container/sub-container.module';
+
+@NgModule({
+    declarations: [FvSectionComponent, FvAccordionItemSectionComponent],
+    imports: [OkFvAccordionItemModule, SubContainerModule],
+    exports: [FvSectionComponent]
+})
+export class FvSectionModule { }