+++ packages/angular-workspace/ok-angular/ex/button/public-api.ts
@@ -1,0 +1,2 @@
+export * from './ok-ex-button.directive';
+export * from './ok-ex-button.module';