+++ packages/ok-components/src/all-components.ts
@@ -6,5 +6,6 @@
 
 import '@ni/spright-components/dist/esm/all-components';
 
-import './button';
+import './fv/all-fv';
+import './ex/all-ex';
 import './icon-dynamic';