+++ packages/react-workspace/spright-react/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/spright-react
 
-<!-- This log was last generated on Wed, 15 Apr 2026 19:52:21 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 20 Apr 2026 22:42:19 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.11.0
+
+Mon, 20 Apr 2026 22:42:19 GMT
+
+### Minor changes
+
+- Add welcome message for spright chat conversation ([ni/nimble@f7e4581](https://github.com/ni/nimble/commit/f7e4581eb131183f6ebccd9f6337b1f54eb9be8f))
+- Bump @ni/spright-components to v6.16.0
+
+## 0.10.33
+
+Mon, 20 Apr 2026 16:30:31 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v6.15.9
 
 ## 0.10.32